requirements.txt
Bootstrap_Flask==2.4.1
Flask_CKEditor==1.0.0
Flask_Login==0.6.3
Flask-Gravatar==0.5.0
Flask_WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.3
Flask==3.1.0
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.38
gunicorn==23.0.0
psycopg2-binary==2.9.10
stripe==11.5.0
email-validator==2.2.0
openai==1.63.2
boto3==1.36.26
pymupdf==1.25.3
pdfplumber==0.11.5
PyPDF2==3.0.1
moviepy==1.0.3
Pillow>=10.2.0
numpy>=1.24.3
moviepy>=1.0.3
#secrets==2.0.0
#MIMEText==0.1.0
#MIMEMultipart==0.1.0